requirements.txt
coverage==4.5.2
coveralls==1.5.1
pillow==5.3.0
python-ev3dev2==2.0.0b2
